+++ src/5gmsaf/generator-5gmsaf
@@ -29,7 +29,7 @@ scriptdir=`cd "$scriptdir"; pwd`
 
 # Command line option defaults
 default_branch='REL-17'
-default_apis="TS26512_M1_ContentHostingProvisioning TS26512_M5_ServiceAccessInformation"
+default_apis="TS26512_M1_ProvisioningSessions TS26512_M1_ContentHostingProvisioning TS26512_M5_ServiceAccessInformation"
 
 # Parse command line arguments
 ARGS=`getopt -n "$scriptname" -o 'a:b:h' -l 'api:,branch:,help' -s sh -- "$@"`
@@ -116,6 +116,8 @@ if [ ! -d "$scriptdir/openapi" ]; then
 fi
 
 "$scriptdir/../../subprojects/rt-common-shared/5gms/scripts/generate_openapi" -a "${APIS}" -b "${BRANCH}" -c "$scriptdir/config.yaml" -l c -d "$scriptdir/openapi" -g 5.2.0 -y "$scriptdir/fix_openapi_yaml.py"
+
+cp set.h openapi/model
 
 exit 0
